fix modern scanner type
