+++ packages/expo-camera/android/src/main/java/expo/modules/camera/analyzers/BarcodeScannerResultSerializer.kt
@@ -12,6 +12,7 @@ object BarCodeScannerResultSerializer {
       putString("data", result.value)
       putString("raw", result.raw)
       putInt("type", result.type)
+      putBundle("extra", result.extra)
       val cornerPointsAndBoundingBox = getCornerPointsAndBoundingBox(result.cornerPoints, result.boundingBox, density)
       putParcelableArrayList("cornerPoints", cornerPointsAndBoundingBox.first)
       putBundle("bounds", cornerPointsAndBoundingBox.second)
@@ -31,7 +32,8 @@ object BarCodeScannerResultSerializer {
       }
     }
 
-    return BarCodeScannerResult(barcode.format, value, raw, cornerPoints, inputImage?.height ?: 0, inputImage?.width ?: 0)
+    val extra = parseExtraDate(barcode)
+    return BarCodeScannerResult(barcode.format, value, raw, extra, cornerPoints, inputImage?.height ?: 0, inputImage?.width ?: 0)
   }
 
   private fun getCornerPointsAndBoundingBox(
@@ -51,6 +53,111 @@ object BarCodeScannerResultSerializer {
       putParcelable("size", getSize(boundingBox.width.toFloat() / density, boundingBox.height.toFloat() / density))
     }
     return Pair(convertedCornerPoints, boundingBoxBundle)
+  }
+
+  fun parseExtraDate(barcode: Barcode): Bundle {
+    val result = Bundle()
+
+    when (barcode.valueType) {
+      Barcode.TYPE_CONTACT_INFO -> {
+        val info = barcode.contactInfo
+        result.apply {
+          putString("type", "contactInfo")
+          putString("firstName", info?.name?.first)
+          putString("middleName", info?.name?.middle)
+          putString("lastName", info?.name?.last)
+          putString("title", info?.title)
+          putString("organization", info?.organization)
+          putString("email", info?.emails?.firstOrNull()?.address)
+          putString("phone", info?.phones?.firstOrNull()?.number)
+          putString("url", info?.urls?.firstOrNull())
+          putString("address", info?.addresses?.firstOrNull()?.addressLines?.firstOrNull())
+        }
+      }
+
+      Barcode.TYPE_GEO -> {
+        val geo = barcode.geoPoint
+        result.apply {
+          putString("type", "geoPoint")
+          putString("lat", geo?.lat.toString())
+          putString("lng", geo?.lng.toString())
+        }
+      }
+
+      Barcode.TYPE_SMS -> {
+        val sms = barcode.sms
+        result.apply {
+          putString("type", "sms")
+          putString("phoneNumber", sms?.phoneNumber)
+          putString("message", sms?.message)
+        }
+      }
+
+      Barcode.TYPE_URL -> {
+        val url = barcode.url
+        result.putString("type", "url")
+        result.putString("url", url?.url)
+      }
+
+      Barcode.TYPE_CALENDAR_EVENT -> {
+        val event = barcode.calendarEvent
+        result.apply {
+          result.putString("type", "calendarEvent")
+          putString("summary", event?.summary)
+          putString("description", event?.description)
+          putString("location", event?.location)
+          putString("start", event?.start?.toString())
+          putString("end", event?.end?.toString())
+        }
+      }
+
+      Barcode.TYPE_DRIVER_LICENSE -> {
+        val license = barcode.driverLicense
+        result.apply {
+          result.putString("type", "driverLicense")
+          putString("firstName", license?.firstName)
+          putString("middleName", license?.middleName)
+          putString("lastName", license?.lastName)
+          putString("licenseNumber", license?.licenseNumber)
+          putString("expiryDate", license?.expiryDate)
+          putString("issueDate", license?.issueDate)
+          putString("addressStreet", license?.addressStreet)
+          putString("addressCity", license?.addressCity)
+          putString("addressState", license?.addressState)
+        }
+      }
+
+      Barcode.TYPE_EMAIL -> {
+        val email = barcode.email
+        result.apply {
+          result.putString("type", "email")
+          putString("address", email?.address)
+          putString("subject", email?.subject)
+          putString("body", email?.body)
+        }
+      }
+
+      Barcode.TYPE_PHONE -> {
+        val phone = barcode.phone
+        result.apply {
+          result.putString("type", "phone")
+          putString("number", phone?.number)
+          putString("phoneNumberType", phone?.type.toString())
+        }
+      }
+
+      Barcode.TYPE_WIFI -> {
+        val wifi = barcode.wifi
+        result.apply {
+          result.putString("type", "wifi")
+          putString("ssid", wifi?.ssid)
+          putString("password", wifi?.password)
+          putString("type", wifi?.encryptionType.toString())
+        }
+      }
+    }
+
+    return result
   }
 
   private fun getSize(width: Float, height: Float) =
